Whitelist image path, moved images to img folder, remove resources folder
