+++ ui/src/main/img/BlackDuckIcon.png
