+++ ui/src/main/img/BlackDuckLogo.png
