+++ ui/src/main/img/BlackDuckLogoAllBlack.png
